kafka-to-chroma-consumer/requirements.txt
# Using CPU-only torch for smaller image size and faster installation
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.3.1
sentence-transformers==2.7.0
chromadb==0.5.0
kafka-python==2.0.2
numpy==1.26.4